requirements.txt
opencv-python
mediapipe
pillow
scikit-learn
pandas
numpy
